feat: [SFI-1617] refactor terminal payment hook
