+++ packages/adyen-retail-react-app/.env.example
@@ -11,6 +11,7 @@ RefArch_ADYEN_LIVE_URL_PREFIX=""
 RefArch_ADYEN_NATIVE_3DS=""
 RefArch_GIFT_CARD_EXPIRATION_TIME=""
 RefArch_ADYEN_POS_ACTIVE_STORE_IDS=""
+RefArch_ADYEN_POS_ENABLED=""
 COMMERCE_API_CLIENT_ID=""
 COMMERCE_API_ORG_ID=""
 COMMERCE_API_SHORT_CODE=""
